pyproject.toml
[tool.poetry]
name = "ct-spiders"
version = "0.1.0"
description = ""
authors = ["Methuselah-Nwodobeh <methuselah.nwodobeh@amalitech.com>"]
readme = "README.md"
packages = [{include = "ct_spiders"}]

[tool.poetry.dependencies]
python = "^3.11"
scrapy = "^2.11.2"
scrapyd = "^1.4.3"
scrapy-prometheus = "^0.4.4"
scrapy-user-agents = "^0.1.1"
scrapyd-client = "^1.2.3"
python-decouple = "^3.8"
boto3 = "^1.34.149"
algoliasearch = "^3.0.0"
requests-aws4auth = "^1.3.1"
pillow = "^10.4.0"
kafka-python = "^2.0.2"
celery = "^5.4.0"
sqlalchemy = "^2.0.31"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
